git log --pretty=format:"%h %ad | %s%d [%an]" merge_branch..android/o/mr1/master:
  cda16e2 Wed Jul 11 15:16:19 2018 +0800 | Fix integer overflow issue in get boottime function (origin/cactus/o_mr1-r0/candidates/mergerequest, origin/android/o/mr1/master, android/o/mr1/master) [Yang, Kai]
  2fdbec9 Fri Jun 29 09:39:56 2018 -0700 | kf4abl: fastboot: do not limit reboot to NORMAL and FASTBOOT [Jeremy Compostella]
  1eb6ab8 Wed Jul 11 09:47:25 2018 +0800 | Avoid reboot again after upgrade bootloader [Zhou, Lihua]
  dc8d849 Wed Jul 4 09:57:53 2018 +0800 | Add the support of show and delete TPM index in userdebug and eng build. [Ming Tan]
  a287bb7 Thu Jul 5 10:25:19 2018 +0800 | Add NULL check for pointer before dereferenced [sunxunou]
  078b075 Thu May 17 21:53:45 2018 +0800 | Add the basic support of USB disk in UEFI based kernelflinger. [Ming Tan]

git diff --stat merge_branch..android/o/mr1/master:
  Android.mk                               |  11 +-
  avb/Android.mk                           |   3 +-
  avb/libavb/avb_cmdline.c                 | 444 ------------------
  avb/libavb/avb_cmdline.h                 |  90 ----
  avb/libavb/avb_crypto.c                  |   3 +-
  avb/libavb/avb_crypto.h                  |   3 -
  avb/libavb/avb_hash_descriptor.c         |   1 -
  avb/libavb/avb_hash_descriptor.h         |  17 +-
  avb/libavb/avb_hashtree_descriptor.c     |   1 -
  avb/libavb/avb_hashtree_descriptor.h     |  17 +-
  avb/libavb/avb_ops.h                     |  85 +---
  avb/libavb/avb_slot_verify.c             | 753 ++++++++++++++++++-------------
  avb/libavb/avb_slot_verify.h             |   6 +-
  avb/libavb/avb_sysdeps.h                 |   4 -
  avb/libavb/avb_sysdeps_posix.c           |   6 -
  avb/libavb/avb_util.c                    |  31 +-
  avb/libavb/avb_util.h                    |  10 -
  avb/libavb/avb_version.h                 |   2 +-
  avb/libavb/uefi_avb_sysdeps.c            |   6 -
  avb/libavb_ab/libavb_ab.h                |  11 -
  include/libkernelflinger/android.h       |   6 +-
  include/libkernelflinger/storage.h       |   3 +
  include/libkernelflinger/timer.h         |   2 +-
  include/libkernelflinger/tpm2_security.h |   5 +
  include/libkernelflinger/version.h       |   2 +-
  kernelflinger.c                          |   6 +-
  kf4abl.c                                 |   7 +-
  libfastboot/fastboot_oem.c               |  46 ++
  libkernelflinger/Android.mk              |   5 +
  libkernelflinger/gpt.c                   |   8 +-
  libkernelflinger/storage.c               |  20 +-
  libkernelflinger/timer.c                 |  14 +-
  libkernelflinger/tpm2_security.c         |  30 ++
  libkernelflinger/trusty_sbl.c            |  94 +---
  libkernelflinger/usb_storage.c           |  73 +++
  libqltipc/ql-tipc/ipc.c                  |   2 -
  libsslsupport/Android.mk                 |  18 +-
  libsslsupport/borningssl/sys/syscall.h   |   2 +-
  38 files changed, 678 insertions(+), 1169 deletions(-)

Tracked-On: https://jira01.devtools.intel.com/browse/OAM-66686

+++ libkernelflinger/timer.c
@@ -75,10 +75,10 @@ __RDTSC (void)
 	return (uint64_t) hi << 32 | lo;
 }
 
-static uint16_t get_cpu_freq(void)
+static uint32_t get_cpu_freq(void)
 {
-	uint16_t cpu_freq;
-	unsigned max_nb_ratio;
+	uint32_t cpu_freq;
+	uint32_t max_nb_ratio;
 	msr_t platform_info;
 
 	platform_info.val = __RDMSR (0xce);
@@ -88,11 +88,11 @@ static uint16_t get_cpu_freq(void)
 	return cpu_freq;
 }
 
-unsigned boottime_in_msec(void)
+uint32_t boottime_in_msec(void)
 {
-	unsigned tick;
-	unsigned bt_us, bt_ms;
-	unsigned cpu_freq;
+	uint64_t tick;
+	uint32_t bt_us, bt_ms;
+	uint32_t cpu_freq;
 
 	cpu_freq = get_cpu_freq();
